Completed the app.

+++ public/client.js
@@ -2,21 +2,26 @@ $( document ).ready(function() {
   var items = [];
   var itemsRaw = [];
   
-  $.getJSON('/api/books', function(data) {
-    //var items = [];
-    itemsRaw = data;
-    $.each(data, function(i, val) {
-      items.push('<li class="bookItem" id="' + i + '">' + val.title + ' - ' + val.commentcount + ' comments</li>');
-      return ( i !== 14 );
+  var getBooks = function getBooks() {
+    $.getJSON('/api/books', function(data) {
+      items = [];
+      itemsRaw = data;
+      $.each(data, function(i, val) {
+        items.push('<li class="bookItem" id="' + i + '">' + val.title + ' - ' + val.commentcount + ' comment' + (val.commentcount !== 1 ? 's' : '') + '</li>');
+        return ( i !== 14 );
+      });
+      if (items.length >= 15) {
+        items.push('<p>...and '+ (data.length - 15)+' more!</p>');
+      }
+      
+      $('#display').html($('<ul/>', {
+        'class': 'listWrapper',
+        html: items.join('')
+      }));
     });
-    if (items.length >= 15) {
-      items.push('<p>...and '+ (data.length - 15)+' more!</p>');
-    }
-    $('<ul/>', {
-      'class': 'listWrapper',
-      html: items.join('')
-      }).appendTo('#display');
-  });
+  };
+
+  getBooks();
   
   var comments = [];
   $('#display').on('click','li.bookItem',function() {
@@ -24,11 +29,11 @@ $( document ).ready(function() {
     $.getJSON('/api/books/'+itemsRaw[this.id]._id, function(data) {
       comments = [];
       $.each(data.comments, function(i, val) {
-        comments.push('<li>' +val+ '</li>');
+        comments.push('<li class="comment">' +val+ '</li>');
       });
-      comments.push('<br><form id="newCommentForm"><input style="width:300px" type="text" class="form-control" id="commentToAdd" name="comment" placeholder="New Comment"></form>');
-      comments.push('<br><button class="btn btn-info addComment" id="'+ data._id+'">Add Comment</button>');
-      comments.push('<button class="btn btn-danger deleteBook" id="'+ data._id+'">Delete Book</button>');
+      comments.push('<form id="newCommentForm" style="margin: 20px 0 0;"><input style="width:300px" type="text" class="form-control" id="commentToAdd" name="comment" placeholder="New Comment" style="margin: 20px 0;"></form>');
+      comments.push('<button class="btn btn-info addComment" id="'+ data._id+'" style="margin: 0 15px 20px 0;">Add Comment</button>');
+      comments.push('<button class="btn btn-danger deleteBook" id="'+ data._id+'" style="margin: 0 0 20px;">Delete Book</button>');
       $('#detailComments').html(comments.join(''));
     });
   });
@@ -39,7 +44,8 @@ $( document ).ready(function() {
       type: 'delete',
       success: function(data) {
         //update list
-        $('#detailComments').html('<p style="color: red;">'+data+'<p><p>Refresh the page</p>');
+        $('#detailComments').html('<p style="color: red;">'+data+'<p><!--<p>Refresh the page</p>-->');
+        getBooks();
       }
     });
   });  
@@ -52,8 +58,9 @@ $( document ).ready(function() {
       dataType: 'json',
       data: $('#newCommentForm').serialize(),
       success: function(data) {
-        comments.unshift(newComment); //adds new comment to top of list
+        comments.unshift('<li class="comment">' + newComment + '</li>'); //adds new comment to top of list
         $('#detailComments').html(comments.join(''));
+        getBooks();
       }
     });
   });
@@ -65,21 +72,29 @@ $( document ).ready(function() {
       dataType: 'json',
       data: $('#newBookForm').serialize(),
       success: function(data) {
-        //update list
+        if (items.length < 14) {
+          $('#display').append('<li class="bookItem" id="' + items.length + '">' + data.title + ' - ' + data.commentcount + ' comment' + (data.commentcount !== 1 ? 's' : '') + '</li>');
+        }
       }
     });
   });
   
   $('#deleteAllBooks').click(function() {
+    var shouldDelete = window.confirm('Are you sure you want to delete all books from this library?');
+    if (!shouldDelete) {
+      return false;
+    }
+
     $.ajax({
       url: '/api/books',
       type: 'delete',
-      dataType: 'json',
+      //dataType: 'json',
       data: $('#newBookForm').serialize(),
       success: function(data) {
-        //update list
+        $('#display').html('');
+        $('#detailComments').html('<p style="color: red;">' + data + '</p>');
       }
     });
-  }); 
-  
+  });
 });
+
